Falcor 5.1 (#296)

+++ Source/Mogwai/Extensions/Capture/FrameCapture.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************
- # Copyright (c) 2015-21, NVIDIA CORPORATION. All rights reserved.
+ # Copyright (c) 2015-22, NVIDIA CORPORATION. All rights reserved.
  #
  # Redistribution and use in source and binary forms, with or without
  # modification, are permitted provided that the following conditions
@@ -175,7 +175,7 @@ namespace Mogwai
             case TextureChannelFlags::RGB: /* No suffix */ outputChannels = 3; break;
             case TextureChannelFlags::RGBA: suffix = ".RGBA"; outputChannels = 4; break;
             default:
-                logWarning("Graph output {} mask {:#x} is not supported. Skipping.", outputName, mask);
+                logWarning("Graph output {} mask {:#x} is not supported. Skipping.", outputName, (uint32_t)mask);
                 continue;
             }
 
@@ -216,7 +216,7 @@ namespace Mogwai
 
                 if (outputFormat == ResourceFormat::Unknown)
                 {
-                    logWarning("Graph output {} mask {:#x} failed to determine output format. Skipping.", outputName, mask);
+                    logWarning("Graph output {} mask {:#x} failed to determine output format. Skipping.", outputName, (uint32_t)mask);
                     continue;
                 }
 
@@ -226,7 +226,7 @@ namespace Mogwai
                 // but then the output image will be in a floating-point format which may be undesirable too.
                 if (is_set(mask, TextureChannelFlags::RGB) && isSrgbFormat(format))
                 {
-                    logWarning("Graph output {} mask {:#x} extracting single RGB channel from SRGB format may lose precision.", outputName, mask);
+                    logWarning("Graph output {} mask {:#x} extracting single RGB channel from SRGB format may lose precision.", outputName, (uint32_t)mask);
                 }
 
                 // Copy color channel into temporary texture.
